Fixed drawing transmission

+++ gui/drawan.h
@@ -2,7 +2,6 @@
 #define DRAWAN_H
 
 void draw_init(void), draw_term(void);
-void draw_brush(GtkWidget *widget,
-		gdouble x, gdouble y, gdouble x2, gdouble y2);
+void draw_brush(GtkWidget *widget, int x, int y, int x2, int y2);
 
 #endif
